+++ scylla/src/transport/mod.rs
@@ -2,7 +2,6 @@ pub mod errors;
 pub mod iterator;
 pub mod legacy_query_result;
 pub mod locator;
-pub(crate) mod metrics;
 pub mod partitioner;
 pub mod query_result;
